requirements.txt
arrow==1.2.3
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==1.7.3
Django==4.1.3
django-cors-headers==3.13.0
django-pwa==1.0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
poyo==0.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==6.1.2
python-webpack-boilerplate==1.0.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.12
whitenoise==6.2.0
